requirement.txt
asgiref==3.4.1
attrs==21.4.0
Django==4.0.1
django-braces==1.15.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.7.0
